Remove uint16 for nonces when calling cSHAKE
cSHAKE implementation takes a uint16 as the nonce or the CSTM, and
the spec was always treating it as a uint16, which required a bunch of
noise in the spec to make sure it didn't overflow or violate other
invariants.

Instead, encapsulate cSHAKE calls into a function that takes nats, which
is what the spec does, and deal with narrowing it to a uint16 in the
function, since the spec implementation we use requires that. Right now
we just assume the cstm fits in a uint16, which realistically will always
be the case but isn't guaranteed since it depends on the RNG not being
broken.

Also remove commented out old assume declaration for gaussSampler.