Update qTESLA spec to v2.3 of 10/31/2018
Some changes to the spec were already incorporated resulting from
discussions with Patrick. Major change is using the hash of the
message as input to hash H, as part of signature generation and
validation, instead of the full message body.

Also, move instantiations of various XOFs and other functions into
parameter sets; we were incorrectly using (c)SHAKE128 in a couple
places we should have been using the -256 versions. Parameterized
functions are now called from the spec.

Manually verified that the spec type-checks across all parameter sets,
and made changes or added refinements where necessary in order for this
to happen.