Extract common definitions into files for heuristic/provable parameter sets
Avoid duplication of common aliases for functions and types in the various
individual parameter sets by having a single set of common ones that vary
only between heuristic and provable, and include it in the per-parameter set
definition file.

Change check_ES to be properly polymorphic for the parameter sets, instead
of only suitable to the provable parameter sets.